[iOS] refactor: multidevice 분리
