+++ iOS/DaDaIkSeon/DaDaIkSeon/Scene/Setting/ViewModel/SettingViewModel+Email.swift
@@ -1,0 +1,29 @@
+//
+//  SettingViewModel+SettingEmail.swift
+//  DaDaIkSeon
+//
+//  Created by 정재명 on 2020/12/13.
+//
+
+import Foundation
+
+// MARK: Email
+extension SettingViewModel {
+    
+      func handlerForEemailSetting(_ input: SettingEmail) {
+          switch input {
+          case .editEmail(let email):
+              if email.count > 5 && email.contains("@") {
+                  state.service.updateEmail(email)
+                  state.email = state.service.readEmail() ?? ""
+                  state.emailEditMode = false
+                  state.emailValidation = true
+              } else {
+                  state.emailValidation = false
+              }
+          case .editEmailMode:
+              state.emailEditMode.toggle()
+              state.emailValidation = true
+          }
+      }
+}
